Move tracing functions to common code
