+++ pkg/channel/consolidated/dispatcher/dispatcher.go
@@ -19,6 +19,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"knative.dev/eventing-kafka/pkg/common/tracing"
 	nethttp "net/http"
 	"strings"
 	"sync"
@@ -179,7 +180,7 @@ func (c consumerMessageHandler) Handle(ctx context.Context, consumerMessage *sar
 		zap.String("subscription", c.sub.String()),
 	)
 
-	ctx, span := startTraceFromMessage(c.logger, ctx, message, consumerMessage.Topic)
+	ctx, span := tracing.StartTraceFromMessage(c.logger, ctx, message, consumerMessage.Topic)
 	defer span.End()
 
 	_, err := c.dispatcher.DispatchMessageWithRetries(
@@ -401,14 +402,4 @@ func (d *KafkaDispatcher) getChannelReferenceFromHost(host string) (eventingchan
 		return cr, eventingchannels.UnknownHostError(host)
 	}
 	return cr, nil
-}
-
-func startTraceFromMessage(logger *zap.SugaredLogger, inCtx context.Context, message *protocolkafka.Message, topic string) (context.Context, *trace.Span) {
-	sc, ok := parseSpanContext(message.Headers)
-	if !ok {
-		logger.Warn("Cannot parse the spancontext, creating a new span")
-		return trace.StartSpan(inCtx, "kafkachannel-"+topic)
-	}
-
-	return trace.StartSpanWithRemoteParent(inCtx, "kafkachannel-"+topic, sc)
 }
